+++ packages/sdks/web-component/CHANGELOG.md
@@ -2,6 +2,50 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [4.1.0](https://github.com/descope/descope-js/compare/web-component-4.0.2...web-component-4.1.0) (2026-08-10)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.24.1`
+* `web-js-sdk` updated to version `1.51.5`
+
+### Features
+
+* **widgets/user-profile-widget:** native bridge for mobile SDKs ([#1415](https://github.com/descope/descope-js/issues/1415)) RELEASE ([0b3fa43](https://github.com/descope/descope-js/commit/0b3fa43eda59a59c4b0938a446bcc844c490a16e))
+
+## [4.0.2](https://github.com/descope/descope-js/compare/web-component-4.0.1...web-component-4.0.2) (2026-08-09)
+
+
+### Bug Fixes
+
+* **web-component:** catch rejected credentials.store promise ([#1457](https://github.com/descope/descope-js/issues/1457)) ([9466aa0](https://github.com/descope/descope-js/commit/9466aa034a07b3ed29e95f2ec6df5da9c3123fcd)), closes [#1456](https://github.com/descope/descope-js/issues/1456)
+
+## [4.0.1](https://github.com/descope/descope-js/compare/web-component-4.0.0...web-component-4.0.1) (2026-08-09)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.24.0`
+* `web-js-sdk` updated to version `1.51.4`
+## [4.0.0](https://github.com/descope/descope-js/compare/web-component-3.70.7...web-component-4.0.0) (2026-08-03)
+
+
+### ⚠ BREAKING CHANGES
+
+* **sdks:** server-produced screen data in onScreenUpdate now lives under context.data instead of directly on context.
+
+### Features
+
+* **sdks:** onScreenUpdate screen data moved under context.data ([#1453](https://github.com/descope/descope-js/issues/1453)) RELEASE ([ddb9dbd](https://github.com/descope/descope-js/commit/ddb9dbd6daa9ae3019c06f3d40d5199ccf3d07ea))
+
+## [3.70.7](https://github.com/descope/descope-js/compare/web-component-3.70.6...web-component-3.70.7) (2026-08-03)
+
+## [3.70.6](https://github.com/descope/descope-js/compare/web-component-3.70.5...web-component-3.70.6) (2026-08-02)
+
+
+### Bug Fixes
+
+* **web-component:** anchor username for password managers on password screens ([#1442](https://github.com/descope/descope-js/issues/1442)) ([db583f3](https://github.com/descope/descope-js/commit/db583f3c71e0635a37a259f37a47cbcb02e2af20)), closes [descope/etc#16712](https://github.com/descope/etc/issues/16712) [descope/etc#16712](https://github.com/descope/etc/issues/16712)
+
 ## [3.70.5](https://github.com/descope/descope-js/compare/web-component-3.70.4...web-component-3.70.5) (2026-07-26)
 
 ### Dependency Updates