+++ packages/widgets/role-management-widget/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.9.11](https://github.com/descope/descope-js/compare/role-management-widget-0.9.10...role-management-widget-0.9.11) (2026-08-10)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.51.5`
+* `sdk-mixins` updated to version `0.24.1`
+* `sdk-component-drivers` updated to version `0.15.0`
+## [0.9.10](https://github.com/descope/descope-js/compare/role-management-widget-0.9.9...role-management-widget-0.9.10) (2026-08-09)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.51.4`
+* `sdk-mixins` updated to version `0.24.0`
 ## [0.9.9](https://github.com/descope/descope-js/compare/role-management-widget-0.9.8...role-management-widget-0.9.9) (2026-07-26)
 
 ### Dependency Updates