+++ packages/widgets/user-management-widget/CHANGELOG.md
@@ -2,6 +2,58 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.17.1](https://github.com/descope/descope-js/compare/user-management-widget-0.17.0...user-management-widget-0.17.1) (2026-08-10)
+
+
+### Bug Fixes
+
+* **user-mgmt-widget:** keep custom-attribute filter columns while schema loads RELEASE ([#1460](https://github.com/descope/descope-js/issues/1460)) ([1a0c793](https://github.com/descope/descope-js/commit/1a0c7934833b8d28440aa851ced0939c9d88305a))
+
+## [0.17.0](https://github.com/descope/descope-js/compare/user-management-widget-0.16.4...user-management-widget-0.17.0) (2026-08-10)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.51.5`
+* `web-component` updated to version `4.1.0`
+* `sdk-mixins` updated to version `0.24.1`
+* `sdk-component-drivers` updated to version `0.15.0`
+
+### Features
+
+* **widgets:** Support UserMgmt Filter ([#1422](https://github.com/descope/descope-js/issues/1422)) ([ce182e9](https://github.com/descope/descope-js/commit/ce182e96c7f29558ad2b38ebb7513beaa030ca7e)), closes [/github.com/descope/etc/issues/14302#issuecomment-5178241391](https://github.com/descope//github.com/descope/etc/issues/14302/issues/issuecomment-5178241391)
+
+## [0.16.4](https://github.com/descope/descope-js/compare/user-management-widget-0.16.3...user-management-widget-0.16.4) (2026-08-09)
+
+### Dependency Updates
+
+* `web-component` updated to version `4.0.2`
+## [0.16.3](https://github.com/descope/descope-js/compare/user-management-widget-0.16.2...user-management-widget-0.16.3) (2026-08-09)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.51.4`
+* `web-component` updated to version `4.0.1`
+* `sdk-mixins` updated to version `0.24.0`
+## [0.16.2](https://github.com/descope/descope-js/compare/user-management-widget-0.16.1...user-management-widget-0.16.2) (2026-08-03)
+
+### Dependency Updates
+
+* `web-component` updated to version `4.0.0`
+## [0.16.1](https://github.com/descope/descope-js/compare/user-management-widget-0.16.0...user-management-widget-0.16.1) (2026-08-03)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.70.7`
+## [0.16.0](https://github.com/descope/descope-js/compare/user-management-widget-0.15.7...user-management-widget-0.16.0) (2026-08-02)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.70.6`
+
+### Features
+
+* **widgets:** recovery email/phone in user management + user profile widgets ([#1445](https://github.com/descope/descope-js/issues/1445)) ([0ea68f8](https://github.com/descope/descope-js/commit/0ea68f875c69ba9a391ff3af282ad1b0d2336256)), closes [descope/backend#1894](https://github.com/descope/backend/issues/1894) [descope/console-app#5598](https://github.com/descope/console-app/issues/5598) [descope/content#2136](https://github.com/descope/content/issues/2136) [descope/etc#15631](https://github.com/descope/etc/issues/15631)
+
 ## [0.15.7](https://github.com/descope/descope-js/compare/user-management-widget-0.15.6...user-management-widget-0.15.7) (2026-07-26)
 
 ### Dependency Updates